requirements.txt
groq
streamlit
mysql-connector-python
pandas
python-dotenv
